+++ export/export.go
@@ -1,4 +1,4 @@
-package resources
+package export
 
 import (
 	"context"
